add ISO 8601 date parsing utility and migrate all event date display/sorting

+++ composeApp/src/wasmJsMain/kotlin/net/winedownwednesday/web/data/Episode.kt
@@ -4,7 +4,7 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class Episode(
-    val id: Int,
+    val id: Long,                       // Long: Date.now() timestamps overflow Int
     val title: String,
     val guestName: String,
     val guestPictureUrl: String,
