+++ composeApp/src/wasmJsMain/kotlin/net/winedownwednesday/web/data/MediaItem.kt
@@ -4,7 +4,7 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class MediaItem(
-    val id: Int,
+    val id: Long,                       // Long: consistent with entity timestamp IDs
     val type: MediaType,
     val thumbnailUrl: String?,
     val contentUrl: String?,
