+++ composeApp/src/wasmJsMain/kotlin/net/winedownwednesday/web/data/models/RSVPRequest.kt
@@ -4,7 +4,7 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class RSVPRequest(
-    val eventId: Int,
+    val eventId: Long,                  // Long: must match Event.id (Date.now() timestamp)
     val firstName: String,
     val lastName: String,
     val email: String,
